обработка ошибок

+++ controllers/cards.js
@@ -1,9 +1,10 @@
 const Card = require('../models/card');
+const { Forbidden } = require('../errors/errors');
 
 const getCards = (req, res) => {
   Card.find({})
     .then((cards) => res.status(200).send(cards))
-    .catch(() => res.status(500).send({ message: 'Произошла ошибка сервера' }));
+    .catch(next);
 };
 
 const createCard = (req, res) => {
@@ -12,10 +13,10 @@ const createCard = (req, res) => {
   Card.create({ name, link, owner })
     .then((card) => res.status(200).send(card))
     .catch((err) => {
-      if (err.name === 'ValidationError') {
-        return res.status(400).send({ message: 'Переданы некорректные данные' });
-      }
-      return res.status(500).send({ message: 'Произошла ошибка сервера' });
+    //   if (err.name === 'ValidationError') {
+    //     return res.status(400).send({ message: 'Переданы некорректные данные' });
+    //   }
+      next(err);
     });
 };
 
@@ -23,22 +24,22 @@ const deleteCard = (req, res) => {
   const id = req.params._id;
   Card.findById(id)
     .then((card) => {
+      if (!card) {
+        throw new NotFound('Данная карточка не найдена');
+      }
       if (card.owner !== req.user._id) {
-        return res.status(403).send({ message: 'Нельзя удалить чужую карточку' });
+        throw new Forbidden('Нельзя удалить чужую карточку');
       }
       return Card.findByIdAndRemove(card._id)
-        .orFail(new Error('NotValidId'))
         .then(() => res.status(200).send({ message: 'Удаление прошло удачно' }))
         .catch((err) => {
-          if (err.name === 'CastError') {
-            return res.status(400).send({ message: 'Переданы некорректные данные' });
-          }
-          if (err.message === 'NotValidId') {
-            return res.status(404).send({ message: 'Данная карточка не найдена' });
-          }
-          return res.status(500).send({ message: 'Произошла ошибка сервера' });
+          // if (err.name === 'CastError') {
+          //   return res.status(400).send({ message: 'Переданы некорректные данные' });
+          // }
+          next(err);
         });
-    });
+    })
+    .catch(next);
 };
 
 const likeCard = (req, res) => {
@@ -48,16 +49,17 @@ const likeCard = (req, res) => {
     { $addToSet: { likes: req.user._id } },
     { new: true },
   )
-    .orFail(new Error('NotValidId'))
-    .then((like) => res.status(200).send(like))
+    .then((like) => {
+      if (!like) {
+        throw new NotFound('Данная карточка не найдена');
+      }
+      return res.status(200).send(like);
+    })
     .catch((err) => {
-      if (err.name === 'CastError') {
-        return res.status(400).send({ message: 'Переданы некорректные данные' });
-      }
-      if (err.message === 'NotValidId') {
-        return res.status(404).send({ message: 'Данная карточка не найдена' });
-      }
-      return res.status(500).send({ message: 'Произошла ошибка сервера' });
+      // if (err.name === 'CastError') {
+      //   return res.status(400).send({ message: 'Переданы некорректные данные' });
+      // }
+      next(err);
     });
 };
 
@@ -68,16 +70,17 @@ const dislikeCard = (req, res) => {
     { $pull: { likes: req.user._id } },
     { new: true },
   )
-    .orFail(new Error('NotValidId'))
-    .then((like) => res.status(200).send(like))
+    .then((like) => {
+      if (!like) {
+        throw new NotFound('Данная карточка не найдена');
+      }
+      return res.status(200).send(like);
+    })
     .catch((err) => {
-      if (err.name === 'CastError') {
-        return res.status(400).send({ message: 'Переданы некорректные данные' });
-      }
-      if (err.message === 'NotValidId') {
-        return res.status(404).send({ message: 'Данная карточка не найдена' });
-      }
-      return res.status(500).send({ message: 'Произошла ошибка сервера' });
+      // if (err.name === 'CastError') {
+      //   return res.status(400).send({ message: 'Переданы некорректные данные' });
+      // }
+      next(err);
     });
 };
